disassemble bytecode

+++ test/Add2.java
@@ -1,13 +1,13 @@
 package tech.icey.r11f.test;
 
 public class Add2 {
-    public static long add(int a, int b) {
+    public static long add(long a, long b) {
         return imp_add(a, b);
     }
 
-    private static long imp_add(int a, int b) {
+    private static long imp_add(long a, long b) {
         return native_add(a, b);
     }
 
-    private static native long native_add(int a, int b);
+    private static native long native_add(long a, long b);
 }
